LibVMI cache : bug fix and unit test
1. fix the bug : key_128_int() misses the high 32bits of the uint64_t low.
As vmi->page_size is a uint32_t type, we can't assume it the same as uint64_t when using the bitwise NOT of it.

2. unit test for LibVMI cache

Thanks to awsaba's help in fixing and unit testing.

+++ libvmi/cache.c
@@ -83,7 +83,7 @@ static gboolean key_128_equals(gconstpointer key1, gconstpointer key2){
  */
 static void key_128_init(vmi_instance_t vmi, key_128_t key, uint64_t low, uint64_t high)
 {
-    low = (low & ~(vmi->page_size - 1));
+    low = (low & ~((uint64_t)vmi->page_size - 1));
     key->low = low;
     key->high = high;
 }
